Better number handling
Numbers are normalized before encoding allowing handling of number big
enough to have a .toString() in sci. notation.
Mantissa and exponent are encoded separatly, exponent is used first for
order, then mantissa (with enough trailing zero to exceed precision)
Nan, +Infinity and -Infinity are handled. Nan is arbitrarly set to be
before -Infinity.